mds_py/pyproject.toml
[tool.poetry]
name = "mds-py"
version = "0.1.0"
description = ""
authors = ["alexmy <alexmy@lisa-park.com>"]
readme = "README.md"
packages = [{include = "mds_py"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
ipython = "^8.7.0"
redis = "^4.3.5"
fastavro = "^1.7.0"
avro = "^1.11.1"
pandas = "^1.5.2"
schema = "^0.7.5"
pyyaml = "^6.0"
polars = "^0.15.2"
yq = "^3.1.0"
strictyaml = "^1.6.2"
ensure = "^1.0.2"
cerberus = "^1.3.4"
ortools = "^9.5.2237"
optional = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"